change according to comments on commit 76fe6fc
